# mypaintings/target/classes/api-spec.yaml
openapi: 3.0.0
info:
  title: MyP@intings
  version: 1.0.0
  description: RESTful API dla aplikacji myp@intings
...
paths:
...
  /ogolne/logowanie:
    post:
      tags:
        - Ogolne
      summary: Logowanie użytkownika
      description: Endpoint do logowania użytkownika do systemu.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_ogolne_logowanie'
      responses:
        '200':
          description: Logowanie zakończone sukcesem
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_logowanie'
        '401':
          description: Nieprawidłowe dane logowania
  /ogolne/rejestracja:
    post:
      tags:
        - Ogolne
      summary: Rejestracja nowego użytkownika
      description: Endpoint do rejestracji nowego użytkownika w systemu.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_ogolne_rejestracja'
      responses:
        '201':
          description: Użytkownik został zarejestrowany pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_rejestracja'
        '400':
          description: Nieprawidłowe dane wejściowe
  /ogolne/support/ticket:
    post:
      tags:
        - Ogolne
      summary: Wyślij ticket do supportu
      description: Endpoint do wysyłania zgłoszenia do działu wsparcia.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_ogolne_support_ticket'
      responses:
        '200':
          description: Ticket został wysłany pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_support_ticket'
        '400':
          description: Nieprawidłowe dane wejściowe
  /ogolne/resolver/{username}:
    get:
      tags:
        - Ogolne
      summary: Uzyskaj ID użytkownika na podstawie nazwy z @
      description: Endpoint pozwala uzyskać ID użytkownika na podstawie nazwy z @.
      parameters:
        - name: username
          in: path
          description: Nazwa użytkownika z @
          required: true
          schema:
            type: string
            example: "@user2024"
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_resolver'
        '404':
          description: Użytkownik nie znaleziony
  /ogolne/avatar/{id}:
    get:
      tags:
        - Ogolne
      summary: Pobierz link do awataru użytkownika
      description: Endpoint zwraca link do obrazu awataru użytkownika o podanym ID.
      parameters:
        - name: id
          in: path
          description: ID użytkownika
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_avatar_get'
        '404':
          description: Użytkownik nie znaleziony
...
  /ogolne/zmiana-hasla:
    put:
      tags:
        - Ogolne
      summary: Zmień hasło użytkownika
      description: Endpoint pozwala zalogowanemu użytkownikowi zmienić swoje hasło.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_ogolne_zmiana_hasla'
      responses:
        '200':
          description: Hasło zostało zmienione pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_zmiana_hasla'
        '400':
          description: Nieprawidłowe dane wejściowe
        '401':
          description: Brak autoryzacji lub nieprawidłowe stare hasło
  /ogolne/dane-dostawy:
    put:
      tags:
        - Ogolne
      summary: Zaktualizuj dane dostawy użytkownika
      description: Endpoint pozwala zalogowanemu użytkownikowi zaktualizować swoje dane dostawy.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_ogolne_dane_dostawy'
      responses:
        '200':
          description: Dane dostawy zostały zaktualizowane pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_dane_dostawy'
        '400':
          description: Nieprawidłowe dane wejściowe
        '401':
          description: Brak autoryzacji
    get:
      tags:
        - Ogolne
      summary: Pobierz dane dostawy użytkownika
      description: Endpoint pozwala zalogowanemu użytkownikowi pobrać swoje dane dostawy.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      responses:
        '200':
          description: Dane dostawy pobrane pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResponseOgolneDaneDostawy'
        '401':
          description: Brak autoryzacji
        '404':
          description: Dane dostawy nie znalezione
  /ogolne/user/nickname/{id}:
    get:
      tags:
        - Ogolne
      summary: Pobierz nick użytkownika z @ na początku
      description: Endpoint zwraca nick użytkownika z @ na początku na podstawie ID użytkownika.
      parameters:
        - name: id
          in: path
          description: ID użytkownika
          required: true
          schema:
            type: integer
            example: 123
      responses:
        '200':
          description: Sukces
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_user_nickname'
        '404':
          description: Użytkownik nie znaleziony
...
  /ogolne/avatar:
    put:
      tags:
        - Ogolne
      summary: Zaktualizuj link do awatara użytkownika
      description: Endpoint pozwala zalogowanemu użytkownikowi zaktualizować link do swojego awatara.
      parameters:
        - name: Authorization
          in: header
          description: Token JWT w formacie "Bearer {token}"
          required: true
          schema:
            type: string
            example: "Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/request_ogolne_avatar_update'
      responses:
        '200':
          description: Awatar został zaktualizowany pomyślnie
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/response_ogolne_avatar_update'
        '400':
          description: Nieprawidłowy URL
        '401':
          description: Brak autoryzacji
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    response_ogolne_logowanie:
      type: object
      properties:
        token:
          type: string
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    response_ogolne_rejestracja:
      type: object
      properties:
        id:
          type: integer
          example: 1
        email:
          type: string
          example: "jan.kowalski@example.com"
        login:
          type: string
          example: "jan_kowalski"
        typ_uzytkownika:
          type: string
          example: "gosc"
    response_ogolne_support_ticket:
      type: object
      properties:
        status:
          type: string
          example: "success"
        message:
          type: string
          example: "Ticket został przyjęty do realizacji."
    response_ogolne_resolver:
      type: object
      properties:
        user_id:
          type: integer
          example: 123
    response_ogolne_avatar_get:
      type: object
      properties:
        avatar_url:
          type: string
          example: "https://example.com/avatars/user123.png"
...
    response_ogolne_zmiana_hasla:
      type: object
      properties:
        message:
          type: string
          example: "Hasło zostało pomyślnie zmienione."
    response_ogolne_dane_dostawy:
      type: object
      properties:
        message:
          type: string
          example: "Dane dostawy zostały zaktualizowane."
        dane_dostawy:
          type: object
          properties:
            street_address:
              type: string
              example: "Baker Street 11/2"
            city:
              type: string
              example: "Szczecin"
            state_province_region:
              type: string
              example: "Zachodniopomorskie"
            zip_code:
              type: string
              example: "70-000"
            email_address:
              type: string
              example: "john_smith@example.com"
            phone_number:
              type: string
              example: "+48 123 456 789"
    response_ogolne_user_nickname:
      type: object
      properties:
        nickname:
          type: string
          example: "@jan_kowalski"
...
    request_ogolne_logowanie:
      type: object
      properties:
        username:
          type: string
          example: "jan_kowalski"
        password:
          type: string
          example: "haslo123"
      required:
        - username
        - password
    request_ogolne_rejestracja:
      type: object
      properties:
        email:
          type: string
          example: "jan.kowalski@example.com"
        login:
          type: string
          example: "jan_kowalski"
        haslo:
          type: string
          example: "haslo123"
        typ_uzytkownika:
          type: string
          enum:
            - gosc
            - artysta
            - hotel
          example: "gosc"
      required:
        - email
        - login
        - haslo
        - typ_uzytkownika
    request_ogolne_support_ticket:
      type: object
      properties:
        imie_i_nazwisko:
          type: string
          example: "Jan Kowalski"
        email:
          type: string
          example: "jan.kowalski@example.com"
        temat:
          type: string
          example: "Problem z rezerwacją"
        wiadomosc:
          type: string
          example: "Nie mogę zalogować się do swojego konta."
      required:
        - imie_i_nazwisko
        - email
        - temat
        - wiadomosc
...
    request_ogolne_zmiana_hasla:
      type: object
      properties:
        stare_haslo:
          type: string
          example: "stareHaslo123"
        nowe_haslo:
          type: string
          example: "noweHaslo456"
      required:
        - stare_haslo
        - nowe_haslo
    request_ogolne_dane_dostawy:
      type: object
      properties:
        street_address:
          type: string
          example: "Baker Street 11/2"
        city:
          type: string
          example: "Szczecin"
        state_province_region:
          type: string
          example: "Zachodniopomorskie"
        zip_code:
          type: string
          example: "70-000"
        email_address:
          type: string
          example: "john_smith@example.com"
        phone_number:
          type: string
          example: "+48 123 456 789"
      required:
        - street_address
        - city
        - state_province_region
        - zip_code
        - email_address
        - phone_number
...
    request_ogolne_avatar_update:
      type: object
      properties:
        avatar_url:
          type: string
          format: uri
          example: "https://example.com/avatars/new_avatar.jpg"
      required:
        - avatar_url
    ResponseOgolneDaneDostawy:
      type: object
      properties:
        message:
          type: string
          example: "Dane dostawy pobrane pomyślnie."
        dane_dostawy:
          type: object
          properties:
            street_address:
              type: string
              example: "Baker Street 11/2"
            city:
              type: string
              example: "Szczecin"
            state_province_region:
              type: string
              example: "Zachodniopomorskie"
            zip_code:
              type: string
              example: "70-000"
            email_address:
              type: string
              example: "john_smith@example.com"
            phone_number:
              type: string
              example: "+48 123 456 789"